move fvcc expression to fvcc/dsl
